ADD seeders for product and category tables through faker + FIX name of pivot table and data type of product's price

+++ database/migrations/2024_09_24_150546_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name', 200);
             $table->text('description');
-            $table->integer('price');
+            $table->decimal('price', 5, 2);
             $table->timestamps();
         });
     }
